rtx: fix moving emissive surfaces light clusters
Do not skip 0th leaf when iterating through R_FatPVS results.

Also:
- move infotool from camera.c to infotool.c
- add hit coords to infotool
- add limited light clusters info into infotool
- add a bit more debug info to light clusters

fixes #279
