+++ ref_vk/vk_light.c
@@ -354,7 +354,7 @@ static void clusterBitMapShutdown( void ) {
 	g_lights_tmp.clusters_bit_map = NULL;
 }
 
-static int computeCellIndex( const int light_cell[3] ) {
+int R_LightCellIndex( const int light_cell[3] ) {
 	if (light_cell[0] < 0 || light_cell[1] < 0 || light_cell[2] < 0
 		|| (light_cell[0] >= g_lights.map.grid_size[0])
 		|| (light_cell[1] >= g_lights.map.grid_size[1])
@@ -403,7 +403,7 @@ vk_light_leaf_set_t *getMapLeafsAffectedByMovingSurface( const msurface_t *surf,
 		gEngine.Con_Reportf("Collecting visible leafs for moving surface %p: %f,%f,%f %f: ", surf,
 			origin[0], origin[1], origin[2], radius);
 
-	for (int i = 1; i <= map->numleafs; ++i) {
+	for (int i = 0; i <= map->numleafs; ++i) {
 		const mleaf_t *leaf = map->leafs + i;
 		if( !CHECKVISBIT( g_lights_bsp.accum.visbytes, i ))
 			continue;
@@ -503,6 +503,10 @@ static qboolean addSurfaceLightToCell( int cell_index, int emissive_surface_inde
 
 	if (cluster->num_emissive_surfaces == MAX_VISIBLE_SURFACE_LIGHTS) {
 		return false;
+	}
+
+	if (debug_dump_lights.enabled) {
+		gEngine.Con_Reportf("    adding surface light %d to cell %d (count=%d)\n", emissive_surface_index, cell_index, cluster->num_emissive_surfaces+1);
 	}
 
 	cluster->emissive_surfaces[cluster->num_emissive_surfaces++] = emissive_surface_index;
@@ -627,11 +631,22 @@ void VK_LightsAddEmissiveSurface( const struct vk_render_geometry_s *geom, const
 			const int min_y = floorf(leaf->minmaxs[1] / LIGHT_GRID_CELL_SIZE);
 			const int min_z = floorf(leaf->minmaxs[2] / LIGHT_GRID_CELL_SIZE);
 
-			const int max_x = ceilf(leaf->minmaxs[3] / LIGHT_GRID_CELL_SIZE);
-			const int max_y = ceilf(leaf->minmaxs[4] / LIGHT_GRID_CELL_SIZE);
-			const int max_z = ceilf(leaf->minmaxs[5] / LIGHT_GRID_CELL_SIZE);
+			const int max_x = floorf(leaf->minmaxs[3] / LIGHT_GRID_CELL_SIZE) + 1;
+			const int max_y = floorf(leaf->minmaxs[4] / LIGHT_GRID_CELL_SIZE) + 1;
+			const int max_z = floorf(leaf->minmaxs[5] / LIGHT_GRID_CELL_SIZE) + 1;
 
-			if (static_map && !canSurfaceLightAffectAABB(world, geom->surf, esurf->emissive, leaf->minmaxs))
+			const qboolean not_visible = static_map && !canSurfaceLightAffectAABB(world, geom->surf, esurf->emissive, leaf->minmaxs);
+
+			if (debug_dump_lights.enabled) {
+				gEngine.Con_Reportf("  adding leaf %d (%d of %d) min=(%d, %d, %d), max=(%d, %d, %d) total=%d\n",
+					leaf->cluster, i, leafs->num,
+					min_x, min_y, min_z,
+					max_x, max_y, max_z,
+					(max_x - min_x) * (max_y - min_y) * (max_z - min_z)
+				);
+			}
+
+			if (not_visible)
 				continue;
 
 			for (int x = min_x; x < max_x; ++x)
@@ -643,7 +658,7 @@ void VK_LightsAddEmissiveSurface( const struct vk_render_geometry_s *geom, const
 					z - g_lights.map.grid_min_cell[2]
 				};
 
-				const int cell_index = computeCellIndex( cell );
+				const int cell_index = R_LightCellIndex( cell );
 				if (cell_index < 0)
 					continue;
 
@@ -694,7 +709,7 @@ static void addLightIndexToleaf( const mleaf_t *leaf, int index ) {
 			z - g_lights.map.grid_min_cell[2]
 		};
 
-		const int cell_index = computeCellIndex( cell );
+		const int cell_index = R_LightCellIndex( cell );
 		if (cell_index < 0)
 			continue;
